Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
APScheduler==3.10.4
bcrypt==4.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
deepface
tensorflow
numpy
Pillow
twilio
sendgrid
psycopg2-binary==2.9.9
requests
